Split out managed implementation
