+++ src/libraries/System.Security.Cryptography/src/System.Security.Cryptography.csproj
@@ -418,6 +418,7 @@
     <Compile Include="System\Security\Cryptography\HashProvider.cs" />
     <Compile Include="System\Security\Cryptography\Helpers.cs" />
     <Compile Include="System\Security\Cryptography\HKDF.cs" />
+    <Compile Include="System\Security\Cryptography\HKDF.Managed.cs" />
     <Compile Include="System\Security\Cryptography\HMAC.cs" />
     <Compile Include="System\Security\Cryptography\HMACCommon.cs" />
     <Compile Include="System\Security\Cryptography\HMACMD5.cs" />
